pyproject.toml
[tool.poetry]
name = "til"
version = "0.1.0"
description = "Keep track of things you learn each day"
authors = ["Kamyar Ghasemlou <github@kamy.me>"]
packages = [
    { include = "til" }
]
[tool.poetry.dependencies]
python = "^3.9"
click = "^7.1.2"
toml = "^0.10.2"
"ruamel.yaml" = "^0.17.4"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
til = "til.main:til"
